refactor(discord): bootstrap の channel config ロードを channel-config.ts へ分離する
