+++ apps/discord/src/bootstrap/channel-config.ts
@@ -1,0 +1,13 @@
+import { existsSync } from "fs";
+import { resolve } from "path";
+
+import { ChannelConfigLoader, type ChannelConfigData } from "../gateway/channel-config-loader.ts";
+
+export async function loadChannelConfig(root: string): Promise<ChannelConfigLoader> {
+	const overlayChannels = resolve(root, "data/context/channels.json");
+	const baseChannels = resolve(root, "context/channels.json");
+	const channelsJson = existsSync(overlayChannels)
+		? await Bun.file(overlayChannels).json()
+		: await Bun.file(baseChannels).json();
+	return new ChannelConfigLoader(channelsJson as ChannelConfigData);
+}
